requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorrt>=8.6.1
numpy>=1.21.0
scikit-learn>=1.0.0
h5py>=3.8.0
pyyaml>=6.0
matplotlib>=3.4.0
tqdm>=4.65.0
pytest>=7.3.1
